1. Implement DI for setting api 2. Remove Base fragment where user api is used

+++ app/src/main/java/com/a2a/app/data/model/EstimateBookingModel.kt
@@ -14,7 +14,7 @@ data class EstimateBookingModel(
         )
 
         data class Pickup(
-            val estimated_price: Int,
+            val estimated_price: Float,
             val pickup_date: String,
             val remarks: String,
             val time: String
